Adding the starting code.
Here in the first line is a comment text and on the second line we have an text which can start or configure the page correctly. And the third line we have the html tag and then head to configure things behind the page and on the next we have the comment again and then the title for the page in the tab area. Next is the comment again then the meta for setting the style of your page. Then again comes a comment. Then the link tag here is used for adding the css and then we have </head> and then </html> to end both of them.